[Zucchini] Fix bugs found by fuzzer in Apply
There was an outstanding TODO to validate that during Apply the writes
to the |new_image| and copies from the |old_image| were within valid
ranges. During raw apply fuzzing a number of memory violations occurred
this is a fix for those.

Bug: 835341
Change-Id: I669304e93e51ba7cd2b862189fbc0a6f3cea1748
Reviewed-on: https://chromium-review.googlesource.com/1028575
Commit-Queue: Calder Kitagawa <ckitagawa@google.com>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#553718}
NOKEYCHECK=True
GitOrigin-RevId: 520aebd16201c0179ff347c7244d940d64af775e