[Zucchini] Require DEX to have parsable code items
Zucchini makes the assumption that a valid DEX file has code items.
However, this contraint was not applied to whether the DEX file
contained valid and parsable code. As a result when attempting to find
references for within these code items, which weren't successfully
parsed, Zucchini would crash.

The solution is to impose a requirement that at least one code item was
parsed to create a disassembler.

Found during fuzzing of DEX files in CL:
https://chromium-review.googlesource.com/c/chromium/src/+/1117123

Change-Id: I76fcbb9267099a7fe3d6eb61c345ffbfaf772fff
Reviewed-on: https://chromium-review.googlesource.com/1118851
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#571276}
NOKEYCHECK=True
GitOrigin-RevId: 2f1a0765a55cda93faa787cf7110db3b78f02a26