[Zucchini] Fix patch size check in fuzzer
The fuzzer found another pathological case of repeated ZTF regions that
causes the patch size check to be violated due to a large number of
headers. The solution should be to increase this upper bound or remove
it entirely. I've gone with the former but if it continues to cause
trouble it might be worth removing as it only finds pathological cases
so far.

Bug: 860070
Change-Id: I276b51bc909ad0da46d9954aee9c98e03dc1973e
Reviewed-on: https://chromium-review.googlesource.com/1126164
Reviewed-by: Samuel Huang <huangs@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Commit-Position: refs/heads/master@{#572581}
NOKEYCHECK=True
GitOrigin-RevId: f45c2a25b3b9b774db04db8d4393727e6a6186f4