[Zucchini] Change DCHECK in reloc_utils.cc to LOG(ERROR)
The (rva_lo_bits & 0xFFF) == rva_lo_bits DCHECK fails when the input
data is invalid. This occurs if the RelocHeader contains |rva_hi_bits|
which are mismatched from the |target_rva| for a given |ref|. This
results in |rva_lo_bits| which are invalid. Until Zucchini is
refactored to support failing to write gracefully we will just print
a loud error and skip the RVA.

This only fails in situations where the input is bad so shouldn't
occur in normal use.

Bug: 860857
Change-Id: I0472286fc54786be26b95f65e6bd18a587bc0c8b
Reviewed-on: https://chromium-review.googlesource.com/1128812
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#573390}
NOKEYCHECK=True
GitOrigin-RevId: e8d1c97b22ece997e5aa5d87d29ef150b6857603