Reformat all gn files in /components/zucchini.
`gn format` recently changed its formatting behavior
for deps, source, and a few other elements when they
are assigned (with =) single-element lists to be consistent
with the formatting of updates (with +=) with single-element.

Now that we've rolled in a GN binary with the change,
reformat all files so that people don't get presubmit
warnings due to this.

This CL was uploaded by git cl split.

R=wfh@chromium.org

Bug: 1041419
Change-Id: Iabb4b3262df63e346520bce93cc00ce6538f76fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1997802
Auto-Submit: Nico Weber <thakis@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#730667}
NOKEYCHECK=True
GitOrigin-RevId: 5366f1719d82185483bbc3fe35725653d384e1ea