[Zucchini] Fix 64-bit ELF reloc code failing to reject references that straddle EOF.
RelocReaderElf is supposed to reject relocs that refers to an abs32
reference that lies outside the image. This should include references
whose offset is in the image, but whose body "straddles" image boundary.
Turns out the reference width for this check was hard-coded to 4, and
on ELF64, this allows 8-byte wide references to sneak by the check (which
leads to check failure down the road).

This CL fixes the above, while adding new unit tests (with refactoring).

Meanwhile, Win32 reloc extraction works a little differently, and
already handles  the above issue.

Bug: 1028892
Change-Id: I746c1a2d8114a429f74d9c0605f50044e05d76be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2013463
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Commit-Queue: Samuel Huang <huangs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#734488}
NOKEYCHECK=True
GitOrigin-RevId: 0b8add5192efb2b283f381fcd9a590e655df2f35