Zucchini: Add the default EXE manifest
This change adds an executable manifest to zucchini.exe so if any of the
code it calls uses Windows version checking functions they work as
intended.

Without a manifest the Windows functions returns version 8.2 for most
version checking APIs and helpers. Refer to the remarks section:
https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-verifyversioninfoa

The lack of a manifest caused zucchini patch generation to hit OOM due
to a recent change in PartitionAlloc which uses Windows version check
functions.

It is better for EXEs to have a compatibility manifest to avoid these
sort of problems and surprises.

Change-Id: I2029f86663dfa4419777494463f55a90b3700acb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3852419
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Commit-Queue: Javier Flores Assad <floresa@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1038813}
NOKEYCHECK=True
GitOrigin-RevId: 5736ae82072fcd980e165eb2494da1ba5f1020f9