Create zucchini64.exe for Windows builds.
32-bit zucchini.exe can run out of memory when constructing a patch.

This CL ensures that there is always a 64-bit zucchini64.exe in the
out directory for both 32-bit and 64-bit builds. For compatibility, a
zucchini.exe also exists, but might be 32-bit or 64-bit depending
on the build configuration.

On 64-bit, the zucchini.exe target is copied to zucchini64.exe.

On 32-bit, a custom target label is created that creates
a 64-bit zucchini.exe using a 64-bit toolchain, then this is
copied to zucchini64.exe in the main outdir.

A future CL will update the builders to always use zucchini64.exe
when constructing diff patches.

BUG=1378898

Change-Id: I66a034307c67b819736001c6a3432401978cc26b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4027674
Reviewed-by: Bruce Dawson <brucedawson@chromium.org>
Commit-Queue: Will Harris <wfh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1071955}
NOKEYCHECK=True
GitOrigin-RevId: 62dfa046f4e16be70a6da96de8324b65cfa243c7