[Zucchini] Fix DCHECK caused by improperly shifted emptied equivalences.
In OffsetMapper, each |equivalences| item maps from a src range to a
dst range, with the same length. Dst ranges don't overlap, and
PruneEquivalencesAndSortBySource() truncates items to remove src range
overlaps. Emptied ranges are removed.

Suppose |current| precedes |reduced| and covers it. To trim |reduced|,
the original logic was:

  delta = current.src_offset + current.length - reduced.src_offset
  reduced.length -= min(reduced.length, delta)
  reduced.src_offset += delta
  reduced.dst_offset += delta

A subtle problem is that if both |delta| and |reduced.dst_offset| are
large, and the |reduced| gets emptied, then |reduced.dst_offset| can
overflow (value won't be unused, but this is still bad). So in
crrev.com/c/4393161 we changed this to:

  delta = current.src_offset + current.length - reduced.src_offset
  capped_delta = min(reduced.length, delta)
  reduced.length -= capped_delta
  reduced.src_offset += capped_delta
  reduced.dst_offset += capped_delta

However, this is causing DCHECK to fail, because if a range gets emptied
then |equivalences| may no longer be sorted by |src_offset|!

This CL fixes the issue by effectively changing the above to:

  reduced.src_offset += delta

This way, when ranges are non-empty, there's no change in behavior; and
when ranges are empty, we'd avoid the |dst_offset| overflow while also
keep |equivalences| sorted by |src_offset| throughout processing.

Bug: 1432457, 1429086
Change-Id: Id61fc0c3466787117ff9d7c5efee66fef8873136
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4418488
Commit-Queue: Etienne Pierre-Doray <etiennep@chromium.org>
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1129897}
NOKEYCHECK=True
GitOrigin-RevId: 4ec735904c32d771e71ebb2954d02bef3d8f0993